+++ source/dsp/inference/utils/HostAudioConfig.h
@@ -5,9 +5,11 @@
 #ifndef NN_INFERENCE_TEMPLATE_HOSTAUDIOCONFIG_H
 #define NN_INFERENCE_TEMPLATE_HOSTAUDIOCONFIG_H
 
+#include <cstddef>
+
 struct HostAudioConfig {
-    int hostChannels;
-    int hostBufferSize;
+    size_t hostChannels;
+    size_t hostBufferSize;
     double hostSampleRate;
 };
